Component vector sin variant, Render ahora controla la vida del nodo de la escena

+++ src/util/GameConstants.hpp
@@ -2,10 +2,10 @@
 
 constexpr std::string_view GAME_NAME	= "Cyborgeddon";
 
-constexpr unsigned UPDATES_PER_SECOND   = 60;
-constexpr float FIXED_DELTA_TIME        = 1.f / UPDATES_PER_SECOND;
+constexpr unsigned  UPDATES_PER_SECOND  = 60;
+constexpr double    FIXED_DELTA_TIME    = 1.0 / UPDATES_PER_SECOND;
 
-constexpr unsigned VIEWPORT_WIDTH       = 1280;
-constexpr unsigned VIEWPORT_HEIGHT      = 720;
+constexpr unsigned  VIEWPORT_WIDTH      = 1280;
+constexpr unsigned  VIEWPORT_HEIGHT     = 720;
 
-constexpr unsigned BACKGROUND_COLOR    	= 0x330044FF;
+constexpr unsigned  BACKGROUND_COLOR    = 0x330044FF;
